RELEASE: v6.01

+++ src/driver/communication/spi/spi0_driver_unittest.c
@@ -28,7 +28,7 @@
 
 #include "cfg_driver_interface.h"
 
-#include "driver/communication/spi/spi0_driver_atmega1284p.h"
+#include "driver/communication/spi/spi0_driver.h"
 #include "local_module_status.h"
 #include "local_mutex.h"
 #include "time_management/time_management.h"
@@ -89,12 +89,12 @@ void spi0_driver_power_off(void) {
 
 }
 
-u8 spi0_driver_bytes_available(void) {
+u16 spi0_driver_bytes_available(void) {
 	return 0;
 }
 
 
-u8 spi0_driver_get_N_bytes(u8 num_bytes, u8* p_buffer_to) {
+u16 spi0_driver_get_N_bytes(u16 num_bytes, u8* p_buffer_to) {
 
 	(void)num_bytes;
 	(void)p_buffer_to;
@@ -103,7 +103,7 @@ u8 spi0_driver_get_N_bytes(u8 num_bytes, u8* p_buffer_to) {
 }
 
 
-u8 spi0_driver_set_N_bytes(u8 num_bytes, const u8* p_buffer_from) {
+u16 spi0_driver_set_N_bytes(u16 num_bytes, const u8* p_buffer_from) {
 
 	(void)num_bytes;
 	(void)p_buffer_from;
@@ -124,7 +124,7 @@ void spi0_driver_start_rx(u16 num_of_rx_bytes) {
 	(void) num_of_rx_bytes;
 }
 
-void spi0_driver_wait_for_rx(u8 num_bytes, u16 timeout_ms) {
+void spi0_driver_wait_for_rx(u16 num_bytes, u16 timeout_ms) {
 	(void) num_bytes;
 	(void) timeout_ms;
 
@@ -138,7 +138,7 @@ void spi0_driver_start_tx(void) {
 
 }
 
-void spi0_driver_wait_for_tx(u8 num_bytes, u16 timeout_ms) {
+void spi0_driver_wait_for_tx(u16 num_bytes, u16 timeout_ms) {
 	(void) num_bytes;
 	(void) timeout_ms;
